chore(spring): Rename file and rebuild file path
